+++ src/view_file_icon.c
@@ -247,7 +247,7 @@ static void vficon_send_layout_select(ViewFile *vf, IconData *id)
 
 		row = g_list_index(vf->list, id);
 		if (row > vficon_index_by_fd(vf, cur_fd) &&
-		    row + 1 < vf_count(vf, NULL))
+		    (guint) (row + 1) < vf_count(vf, NULL))
 			{
 			read_ahead_fd = vf_index_get_data(vf, row + 1);
 			}
@@ -809,7 +809,7 @@ gint vficon_index_is_selected(ViewFile *vf, gint row)
 	return (id->selected & SELECTION_SELECTED);
 }
 
-gint vficon_selection_count(ViewFile *vf, gint64 *bytes)
+guint vficon_selection_count(ViewFile *vf, gint64 *bytes)
 {
 	if (bytes)
 		{
@@ -1970,7 +1970,7 @@ static gint vficon_index_by_id(ViewFile *vf, IconData *in_id)
 	return -1;
 }
 
-gint vficon_count(ViewFile *vf, gint64 *bytes)
+guint vficon_count(ViewFile *vf, gint64 *bytes)
 {
 	if (bytes)
 		{
